// src/Services/interface.ts
export interface Product {
    amount: number;
    ownerName: string;
    productName: string;
    description: string;
    imageUrl: string;
    recommendedImages: string[];
    title: string;
    subtitle: string;
    ownerImage:string
    isActive: boolean,
    isOutOfStock: boolean,
    createdAt: any
    id:string
  }
  